do not implicitly fallback to legacy implementation
Summary:
as $title

the new `threaded-all-reduce` algorithm has few requirements to run: such as all ranks have to be threaded, or count (element) has to meet current platform's requirement. if not, we implicitly fallback to legacy all-reduce implementation. This implicit fallback behavior makes developer hard to know if the new algorithm is working or not.

changes in this diff:
- be explicit. if `threaded` is specified by user, but can not support, just return ncclError as is. no fallback
- increase MIN_COUNT to meet threaded-all-reduce requirements
- fixed a bug to determine devId

> NOTE: we can bring the fallback behavior once this new algorithm gets matured and are ready for prod use, that time we can just drop the knob all together and let it pick the optimal one

Reviewed By: wesbland

Differential Revision: D46993775

fbshipit-source-id: 5ac5ff4036dcca787fae6940416bb6d50f837e72