Simplify the DDA code structure
Summary:
This patch allows each rank to maintain a copy of the metadata,
instead of trying to share them between multiple ranks.  The metadata
does not take too much memory (a few KB) and the complexity associated
with the sharing is unnecessary, especially for the HCM case.

In this new code structure, we simply pass a pointer to the full
copy of the global metadata to the kernel, and let the kernel figure
out what information it needs to access.  This also simplifies the
thread-shared data structure, which now only tracks the list of
registered ranks.

Reviewed By: cenzhaometa

Differential Revision: D48593639

fbshipit-source-id: 876d395556b862a34c3d4057c2c7ee900bfa8a50