print ctran allgather info only when ctran algo is used
Summary:
Previous code log collective info whenevern allgather is called, no matter ctran algorithm is used or not. It causes duplicated log when default allgather is called, and also increased opCount incorrectly in such a case.

This patch fixes it.

In addition, this patch also reformat the log so that it matches default collective path's log format (see log example in test plan)

Reviewed By: kingchc

Differential Revision: D50251809

fbshipit-source-id: 560d8bccbeecb3031c066480aaaa3c9c0e9f2eea