Initial version of QP scaling
Summary:
Allow core data communication to go over multiple QPs.  The intent is
to scale the number of QPs based on message size and communicator
size.  The intent is to create many flows, so the network fabric can
spray data across this links as effectively as possible.  But we do
not want to make the size of each message very small either.

Reviewed By: minsii

Differential Revision: D50634493

fbshipit-source-id: e8a8432f7b99aac5cc67ac5f353ed8b68f745bd9