Maintain singleton class for IB context/pd
Summary:
We don't need to create a separate context and PD for each
communicator.  We can simply maintain a singleton class for it.

One shortcoming of this model is that we now have to create context/pd
for each device (although no QPs or CQs are created for each device).

Reviewed By: steinbrecher

Differential Revision: D50781398

fbshipit-source-id: cba7c96d8f6e193d1e04843e6ce350e28e18a701