Bug-fix for the QP scaling code
Summary:
QP scaling relies on the receiver process being able to calculate the
number of QPs used based on the message size.  Unfortunately, instead
of adding the message size to the immediate field, we had added it to
the wr_id, which is local and not visible to the remote process.

Reviewed By: kingchc

Differential Revision: D50827895

fbshipit-source-id: 952fd7d1830c2c8fc5e1dcbbdb3dd73f62141605