fix lazy registration + dynamic registration bug and refactor
Summary:
The previous code fails when enable lazy registration (default) and hit a buffer that is not preregistered by user (e.g., if we disable Pytorch registration hook).

To fix this issue, we have to refactor the the code a bit to handle all possible cases:
1. Disabled registration hook in Pytorch, all buffers are dynamically registered at collective time
2. Enabled registration hook and lazy registration, all buffers are cached and partial buffers are registered at first collective
3. Enable registration hook and early registration, all buffers are pre-registered before collectives.

In this patch, we also refactored the registration counters to print more details. This would be needed to figure out the bug and useful for troubleshooting/perf understanding. We expect frequency of registration logs should be at most per collective, thus setting to NCCL_COLL.

Reviewed By: kingchc

Differential Revision: D50958307

fbshipit-source-id: 1e2050099dd6ea9ae7bb9ab4e7349efb60db0263