Parser for control variables
Summary:
Builds a new yaml based model for declaring control variables as a
replacement to the NCCL_PARAM model.  This method allows more control
on what the control variables represent (e.g., integers, strings,
enums), the choices available to the user, documentation on what the
control variable does, etc.

The code also looks for environment variables in the NCCL_ namespace
that are not declared by NCCL to avoid namespace pollution.

Reviewed By: wesbland

Differential Revision: D51014921

fbshipit-source-id: 3822c325609c6653ffd8f095287af5d66c92cfe4