Add CTRAN topo file parsing + UT
Summary:
Need to know what zone things are in for cross-zone. Extracts this info from an arbitrary file in `KEY=VALUE` format, with the list of keys supplied in a second environment variable.

This assumes that topology is hierarchical by host, which won't apply for rail-optimized topologies, but we can probably resolve that with, say, a list of values in the VALUE for all the GPUs in the host.

For use with fbwhoami:
```
NCCL_CTRAN_TOPO_FILE=/etc/fbwhoami
NCCL_CTRAN_TOPO_FILE_KEYS=DEVICE_TC_NETWORK_ID,DEVICE_RACK_SERIAL,DEVICE_ID
```

NOTE: Right now, we assume the values are integers. Definitely not the "right" thing to do but works for now :)

Reviewed By: minsii

Differential Revision: D51895742

fbshipit-source-id: aeabf3fd2c414e99f4d6774d616d4603a5225b9e