Remove NVS tag for NVS-only algorithms
Summary:
There is no HCM version for some algorithms, so there is
no need to tag the existing version with _NVS.

Even if we do develop a non-NVS version for these algorithms, there is
no need to tag the platform information to the env variable.  We can
simply give different values to the same variable depending on the
platform.

Reviewed By: cenzhaometa

Differential Revision: D52427192

fbshipit-source-id: 5294041a336bb8082c5ccb156afb5c0131cc65dc