add ncclAllToAllv API
Summary:
It would help ctran to optimize alltoall if it is passed in via a specific API, rather than the original grouped send/recv based approach. This patch adds the API which internally calls grouped send/recv, same as the original approach.

Ctran alltoall will be introduced in later diff.

Reviewed By: kingchc

Differential Revision: D52335580

fbshipit-source-id: aa9cdd109bc9e01b68b19898833d7972347d9792