Separate out threaded and IPC barrier flags
Summary:
The threaded barrier flag is restrictive in that it can take only the
values '0' or '1' (so it can fit in a single bit at the end of the
2-byte aligned address).  The IPC barrier flag, on the other hand,
does not have this restriction because it's not combined with the
address exchange.

This patch separates out the threaded and IPC barrier flags, so we can
use a simple monotonically increasing value for the IPC barrier flag,
thus reducing the number of mailboxes that we need to maintain.

Reviewed By: cenzhaometa

Differential Revision: D52426179

fbshipit-source-id: 26718fe0397d6f6978ca088b1ff0b629bd59043e