add macro for check with exception (#40)
Summary:
Pull Request resolved: https://github.com/facebookresearch/nccl/pull/40

With previous approach using `CUDACHECKGOTO`|`NCCLCHECKGOTO`, it required `ncclResult_t res` variable and a goto label. C++ compiler complains `res` is set but not used.

It is more straightforward to just throw the exception within the CHECK macro.

This patch defines `CUDACHECKTHROW` and `NCCLCHECKTHROW` macros for error checking in ctran constructors.

Reviewed By: cenzhaometa

Differential Revision: D53187428

fbshipit-source-id: 5ddc56643740419c35f023e750c4f3d5561ac76e